# cuda4dnn(mish): use fp32 mish for fp16 mish (faster and more accurate)