--- a/modules/dnn/src/cuda/functors.hpp
+++ b/modules/dnn/src/cuda/functors.hpp
@@ -57,10 +57,18 @@ struct mish_functor<float> {
         auto n = e * e + 2 * e;
         if (value <= -0.6f)
             return value * fast_divide(n, n + 2);
-
         return value - 2 * fast_divide(value, n + 2);
     }
 };
+
+#if !defined(__CUDA_ARCH__) || (__CUDA_ARCH__ >= 530)
+template <>
+struct mish_functor<__half> {
+    __device__ __half operator()(__half value) {
+        return mish_functor<float>()(value);
+    }
+};
+#endif
 
 template <class T>
 struct sigmoid_functor {